% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdt.R
\name{asdt}
\alias{asdt}
\title{Associated single decrement life table (ASDT) for causes of death (cause-deleted life table)}
\usage{
asdt(age, m_all, m_i, full = FALSE, method = "chiang1968", ...)
}
\arguments{
\item{age}{Numeric array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{m_all}{Numeric array with age specific mortality rates of all causes of death (usual mx).}

\item{m_i}{Numeric array with age specific mortality rates of some cause of death (i)}

\item{full}{Logical. Is full table needed? \code{TRUE} for full, \code{FALSE} for concise. By default, \code{FALSE}}

\item{method}{Character. The method of ASDT construction to use. Now just "chiang1968" is supported.}

\item{...}{Optional. Additional arguments for \code{\link[=LT]{LT()}} function.}
}
\value{
dataframe.
}
\description{
Associated single decrement life table (ASDT) for causes of death (cause-deleted life table)
}
\references{
Chiang, L. (1968). \emph{Introduction to Stochastic Processes in Biostatistics}. New York: John Wiley and Sons.
}
