% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_chunk.R
\name{code_from_hook}
\alias{code_from_hook}
\title{Converts raw chunk to a string of code}
\usage{
code_from_hook(.chunk)
}
\description{
Raw chunk string has been pulled from within the \code{"source"} hook using \code{knitr::knit_hooks$get()}.  Backticks and trailing/leading lines are stripped, all formatting is otherwise perserved.
}
