% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo.R
\name{create_demo}
\alias{create_demo}
\title{Builds a demo_code object from a code chunk}
\usage{
create_demo(label)
}
\arguments{
\item{label}{creates demo code object from a}
}
\value{
An object of class \code{\link{demo_code}}
}
\description{
This function reads the source code from a given code chunk that has the value \code{label} set to the \code{demo} option; i.e., \code{{r, demo = label}}.
}
\details{
When run directly in a source file, \code{create_demo()} reads the text of the active file and extracts the relevant string of source code.  (Important: this only works in RStudio!)

When run during the \code{knitr::knit()} process, \code{create_demo()} pulls the relevant chunk source during \code{knitr::knit_hooks$set("source").}
}
