% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_chunk.R
\name{demo_chunk}
\alias{demo_chunk}
\title{Builds a \code{\link{demo_code}} object from a code chunk}
\usage{
demo_chunk(label)
}
\arguments{
\item{label}{String that gives the name of the label used in a chunk option.}
}
\value{
An object of class \code{\link{demo_code}}
}
\description{
This function reads the source code from a given code chunk that has the value \code{label} set to the \code{demo} option; i.e., \code{{r, demo = "my_label"}}.
}
\details{
When run directly in a source file, \code{demo_chunk()} reads the text of the active file and extracts the relevant string of source code from the labelled chunk.  (Important: this only works in RStudio!)

When run during the \code{knitr::knit()} process, \code{demo_chunk()} pulls the relevant chunk source during \code{knitr::knit_hooks$set("source").}
}
