% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_code.R
\name{demo_code}
\alias{demo_code}
\title{Creates an object of the class \code{demo_code}}
\usage{
demo_code(.code_string, eval_here = TRUE)
}
\arguments{
\item{.code_string}{A string containing executable R code.}

\item{eval_here}{A boolean specifying whether the code should be immediately evaluated, in addition to creating the \code{demo_code} object. (Defaults to \code{TRUE})}
}
\value{
A \code{demo_code} object.
}
\description{
\code{demo_code} objects are evaluated R code, returned from \code{evaluate::evaluate}, with an attached attribute called \code{print_string} which sets up fancy formatting for knitting.
}
\examples{

# When run in console, this will print only the results of mean(1:10)
my_dc <- demo_code('mean(1:10)') \%>\% hlt_funs()

# The demo_code object itself has no output

my_dc

# However, when knitted, the source code is formatted.

attr(my_dc, "print_string")


# Objects defined in demo_code are created in the environment

demo_code('foo <- mean(1:10)')

foo + 5



}
\seealso{
\code{\link{hlt_*}}
}
