% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demor-package.R
\docType{package}
\name{demor-package}
\alias{demor}
\alias{demor-package}
\title{demor: Methods for Demographic Analysis}
\description{
Implements life tables, fertility and mortality indicators, decomposition methods, Lee-Carter mortality forecasting, Leslie matrices, and population pyramids for demographic analysis. Methods are described in Preston et al. (2001, ISBN:1557864519) and Ustyuzhanin (2025) \doi{10.17323/demreview.v12i4.30415}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vadvu.github.io/demor/}
  \item \url{https://github.com/vadvu/demor}
  \item Report bugs at \url{https://github.com/vadvu/demor/issues}
}

}
\author{
\strong{Maintainer}: Vadim Ustyuzhanin \email{vadim.v.ustyuzhanin@gmail.com} (\href{https://orcid.org/0000-0003-3800-1108}{ORCID})

}
\keyword{internal}
