% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoR.R
\docType{package}
\name{demoR}
\alias{demoR}
\alias{demoR-package}
\title{demoR: A package for clear demonstration of R code.}
\description{
The demoR package provides ...
}
\section{Highlight functions}{


The \code{hlt_*} functions add formatting to printed code.
}

\section{Demo functions}{


The \code{demo_*} functions print and run the code.
}

