% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udfs.R
\name{demor_ht}
\alias{demor_ht}
\title{Display some head/tail records}
\usage{
demor_ht(df, N1 = 5, N2 = 5)
}
\arguments{
\item{df}{An input data frame.}

\item{N1}{Number of head to show; default 5.}

\item{N2}{Number of tail to show; default 5.}
}
\value{
A data frame of some head and tail of the input data frame.
}
\description{
Display some head/tail records
}
\examples{
demor_ht(mtcars)
}
