% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Gini coefficient of a life table}
\usage{
gini(age, mx, ...)
}
\arguments{
\item{age}{Numeric array of age intervals}

\item{mx}{Numeric array with age specific mortality rates.}

\item{...}{Optional. Additional arguments for \code{\link[=LT]{LT()}} function.}
}
\value{
A list with two components: \code{Gini}, itself a list with the relative
Gini coefficient \code{G0} and the absolute Gini coefficient \code{G0_abs}; and
\code{plot}, a data frame with columns \code{Fx} and \code{Phix} for drawing the Lorenz
curve.
}
\description{
Gini coefficient of a life table
}
\examples{
age <- 0:5
mx <- c(0.02, 0.01, 0.012, 0.015, 0.02, 0.03)
gini(age, mx)$Gini
}
