% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{hlt_regexp.demo_code}
\alias{hlt_regexp.demo_code}
\title{S3 method for \code{\link{demo_code}} objects}
\usage{
\method{hlt_regexp}{demo_code}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{demo_code}}.}
}
\value{
An object of class \code{\link{demo_code}}.
}
\description{
Applies highlighting to the \code{print_string} attribute.
}
