% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{hlt_regexp}
\alias{hlt_regexp}
\title{Highlights parts of a string}
\usage{
hlt_regexp(.string, pattern, code = TRUE, ...)
}
\arguments{
\item{.string}{A string object}

\item{pattern}{A regular expression to match}

\item{code}{Should this string be displayed in R code format?}

\item{...}{Formatting options, passed to \code{\link{txt_style}}}

\item{hlt_color}{Color to highlight code with.  Defaults to}
}
\description{
\code{hlt_*} returns an string of R code with formatting wrappers (currently only html)
}
