% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udfs.R
\name{htnames}
\alias{htnames}
\title{Displays head, tail, and column names of a data frame}
\usage{
htnames(df)
}
\arguments{
\item{df}{An input data frame.}
}
\value{
A list containing head, tail, and column names of the input data frame df.
}
\description{
Displays head, tail, and column names of a data frame
}
\examples{
htnames(mtcars)
}
