% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie.R
\name{leslie}
\alias{leslie}
\title{Leslie Matrix}
\usage{
leslie(mx, fx, age.mx, age.fx, srb = 100/205, fin = TRUE, ...)
}
\arguments{
\item{mx}{Numeric array of age specific mortality rates.}

\item{fx}{Numeric array of age specific fertility rates.}

\item{age.mx}{Numeric array of ages for mx.}

\item{age.fx}{Numeric array of ages for fx.}

\item{srb}{Numeric. Sex ratio at birth. Usually it is assumed that for males it is \code{105/205}, for females it is \code{100/205}. By default, it is \code{100/205}.}

\item{fin}{Logical. Should the survival rate for the last age-group be nonzero? By default it is \code{FALSE}, so the last survival rate is 0 as in classical model. Otherwise, it is \eqn{T_{x}/T_{x-1}}.}

\item{...}{Optional. Additional arguments for \code{\link[=LT]{LT()}} function.}
}
\value{
A square numeric matrix of class \code{leslie} with one row and one
column per age group. The first row contains fertility contributions and
the subdiagonal contains survival ratios.
}
\description{
Leslie Matrix
}
\examples{
mx <- c(0.02, 0.01, 0.012, 0.015, 0.02)
fx <- c(0.05, 0.08)
leslie(mx = mx, fx = fx, age.mx = 0:4, age.fx = 1:2)
}
\seealso{
\code{\link[=summary.leslie]{summary.leslie()}} for \code{leslie} output that calculates \eqn{\lambda}, \eqn{r}, \eqn{w} and \eqn{v}.
}
