% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT.R
\name{LT}
\alias{LT}
\title{Life table}
\usage{
LT(age, sex, mx)
}
\arguments{
\item{age}{Array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{sex}{Sex. "m" for males or "f" for females.}

\item{mx}{Age specific mortality rates.}
}
\value{
Matrix
}
\description{
Life table
}
\details{
ax for age 0 is modeled as in Andreev & Kingkade (2015).
}
\references{
Andreev, E. M., & Kingkade, W. W. (2015). Average age at death in infancy and infant mortality level: Reconsidering the Coale-Demeny formulas at current levels of low mortality. \emph{Demographic Research}, \emph{33}, 363-390.
}
