% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{mask}
\alias{mask}
\alias{mask_regexp}
\title{Blanks out part of the code}
\usage{
mask(.string, pattern, ...)

mask_regexp(.string, pattern, ...)
}
\arguments{
\item{.string}{A string object}

\item{pattern}{A regular expression to match}

\item{...}{Further formatting options, passed to \code{\link{txt_style}}}
}
\description{
Blanks out part of the code
}
