% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.age.R
\name{med.age}
\alias{med.age}
\title{Median age calculation}
\usage{
med.age(N, age)
}
\arguments{
\item{N}{Numeric array. Population counts by age groups (from young to old)}

\item{age}{Numeric array. Lower bounds of age groups, same length as \code{N}}
}
\value{
A length-1 numeric value giving the estimated median age of the
population represented by \code{N}.
}
\description{
Median age calculation
}
\examples{
N <- c(100, 90, 80, 70, 60)
age <- seq(0, 20, 5)
med.age(N, age)
}
