% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLT.R
\name{MLT}
\alias{MLT}
\title{Multiple Decrement Life Table}
\usage{
MLT(age, mx, ...)
}
\arguments{
\item{age}{Numeric array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{mx}{List of numeric arrays. 1st array should be all-cause mx in the population, other arrays are cause-specific mx in the population.}

\item{...}{Other parameters for the function \code{\link[=LT]{LT()}} including \code{ax} (by default, the middle of the interval), \code{sex} (by default = \code{"m"} - males), \code{l0} (by default = 1).}
}
\value{
A numeric matrix extending the output of \code{\link[=LT]{LT()}}. In addition to the
standard life-table columns, for each cause \code{i} in \code{mx[-1]} it adds
\code{qx_i}, \code{dx_i}, \code{lx_i}, and \code{ex_no_i}, corresponding to cause-specific
death probabilities, deaths, survivors, and cause-deleted life expectancy.
}
\description{
Multiple Decrement Life Table
}
\examples{
data(asdtex)
mx_causes <- list(
  all = asdtex$all,
  neoplasms = asdtex$neoplasms,
  circulatory = asdtex$circulatory
)
MLT(age = asdtex$age, mx = mx_causes)[1:3, c("age", "mx", "qx_neoplasms", "ex_no_neoplasms")]
}
\seealso{
\code{\link[=LT]{LT()}} for usual life table calculation
}
