% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm.table.R
\name{msm.table}
\alias{msm.table}
\title{Multi-state life table based on multi-state markov model from msm package}
\usage{
msm.table(model, ages, start, covs)
}
\arguments{
\item{model}{\code{msm} model. See details.}

\item{ages}{Numeric vector of ages.}

\item{start}{Start values of lx for each group presented as numeric vector.}

\item{covs}{List of (age-invariant) covariates as for \code{msm} models.}
}
\value{
Table with age (x), mx (transition rates to death), lx (root of the table = 1), Lx, Tx and ex for all groups and the whole population.
}
\description{
Multi-state life table based on multi-state markov model from msm package
}
\details{
\code{msm} is a Multi-state Markov model in continuous time, where age is both time and covariate.
For example, \code{model = msm::msm(state~age, subject = id, covariates = ~ age + sex, data = data, ...)}
}
\examples{
# model <- msm::msm(state~age, subject = id, covariates = ~ age + sex, data = data, ...)
# table <- msm.table(model = model, ages = 20:80, start = c(1,0,0), covs = list(sex = 1))

}
