% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_code.R
\name{print.demo_code}
\alias{print.demo_code}
\title{S3 method for printing a \code{demo_code}}
\usage{
\method{print}{demo_code}(x, ...)
}
\description{
Print results of evaluating sources.  Do NOT print any extra \code{demo_code} object info.
}
