% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.proj.R
\name{prob.proj}
\alias{prob.proj}
\title{Probabilistic Projections of Population}
\usage{
prob.proj(mort.m, mort.f, fert, time.n, sims, popf, popm)
}
\arguments{
\item{mort.m}{Data.frame with columns year, age, mx for males.}

\item{mort.f}{Data.frame with columns year, age, mx for females}

\item{fert}{Data.frame with columns year, age, fx.}

\item{time.n}{Numeric value of for how many years projection should be done.}

\item{sims}{Numeric values of number of simulations.}

\item{popf}{Data.frame with columns year, age, pop for females}

\item{popm}{Data.frame with columns year, age, pop for females}
}
\value{
list with data.frame with projections and VAR model
}
\description{
Probabilistic Projections of Population
}
