% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosbris.codes}
\alias{rosbris.codes}
\title{Rosbris' Region codes}
\format{
\subsection{A dataframe}{

\describe{
\item{n}{number of Region (sequence number)}
\item{name}{Region name}
\item{code}{Region unique code}
}
}
}
\source{
Rosbris \href{http://demogr.nes.ru/index.php/ru/demogr_indicat/data}{webpage}
}
\usage{
rosbris.codes
}
\description{
Rosbris' Region codes
}
\keyword{datasets}
