% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosbris_fertility_pop_1}
\alias{rosbris_fertility_pop_1}
\title{Rosbris' Fertility and population (1-year interval)}
\format{
\subsection{A dataframe}{

\describe{
\item{year}{year}
\item{code}{number of Region}
\item{territory}{rural/urban/total}
\item{age}{age}
\item{fx}{ASFR}
\item{N}{Population}
\item{Bx}{Number of Births}
\item{name}{Region name}
}
}
}
\source{
Russian Fertility and Mortality Database. Center for Demographic Research, Moscow (Russia). Available at https://www.nes.ru/demogr-fermort-data
}
\usage{
rosbris_fertility_pop_1
}
\description{
Rosbris' Fertility and population (1-year interval)
}
\keyword{datasets}
