% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosbris_fertility_pop_5}
\alias{rosbris_fertility_pop_5}
\title{Rosbris' Fertility and population (5-year interval)}
\format{
\subsection{A dataframe}{

\describe{
\item{year}{year}
\item{code}{number of Region}
\item{territory}{rural/urban/total}
\item{age}{age}
\item{fx}{ASFR}
\item{fx1}{ASFR, 1 order}
\item{fx2}{ASFR, 2 order}
\item{fx3}{ASFR, 3 order}
\item{fx4}{ASFR, 4 order}
\item{fx5}{ASFR, 5+ order}
\item{N}{Population}
\item{Bx}{Number of Births}
\item{Bx1}{Number of Births, 1 order}
\item{Bx2}{Number of Births, 2 order}
\item{Bx3}{Number of Births, 3 order}
\item{Bx4}{Number of Births, 4 order}
\item{Bx5}{Number of Births, 5+ order}
\item{name}{Region name}
}
}
}
\source{
Rosbris \href{http://demogr.nes.ru/index.php/ru/demogr_indicat/data}{webpage}
}
\usage{
rosbris_fertility_pop_5
}
\description{
Rosbris' Fertility and population (5-year interval)
}
\keyword{datasets}
