% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sle_stand}
\alias{sle_stand}
\title{Data on Standard Life expectancies that is used for YLL calculations}
\format{
\subsection{A dataframe}{

\describe{
\item{stand}{Standard: 1 - World Health Organization Standard Life Expectancy by single-age; 2 - Global Burden of Disease studies (GBD) and WHO Global Health Estimates (WHO GHE) Standard Life Expectancy by 5-year age groups.}
\item{age}{age}
\item{ex}{Standard Life Expectancy}
}
}
}
\source{
Martinez, R., Soliz, P., Caixeta, R., Ordunez, P. \href{https://doi.org/10.1093/ije/dyy254}{2019}. Reflection on modern methods: years of life lost due to premature mortality—a versatile and comprehensive measure for monitoring non-communicable disease mortality. International Journal of Epidemiology 48, 1367–1376.
}
\usage{
sle_stand
}
\description{
Data on Standard Life expectancies that is used for YLL calculations
}
\keyword{datasets}
