% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr.R
\name{tfr}
\alias{tfr}
\title{Total Fertility Rate (TFR)}
\usage{
tfr(fx, age.int = 1)
}
\arguments{
\item{fx}{Numeric array of age specific fertility rates.}

\item{age.int}{Numeric. Age group: \code{1} for one-year, \code{5} for five-year. Any age groups are allowed.}
}
\value{
A length-1 numeric value equal to the sum of age-specific fertility
rates multiplied by \code{age.int}.
}
\description{
Total Fertility Rate (TFR)
}
\examples{
fx <- c(0.02, 0.08, 0.12, 0.09, 0.04, 0.01, 0.001)
tfr(fx, age.int = 5)
}
