% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_style.R
\name{txt_style}
\alias{txt_style}
\title{Wraps text in html or latex code for formatting}
\usage{
txt_style(x, type = "html", bold = FALSE, underline = FALSE,
  italics = FALSE, ...)
}
\arguments{
\item{x}{The string to be wrapped}

\item{type}{The style of display, defaults to "html"  (currently nothing else is supported, sorry)}

\item{bold}{Should the text be bolded?}

\item{underline}{Should the text be underlined?}

\item{italics}{Should the text be italicized?}

\item{...}{various display options: any html CSS \code{style} options, or one of \code{font}, \code{size}, \code{color}, \code{background}, \code{style}.}
}
\description{
\code{txt_style} returns a string including styling wrappers
}
