% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ages.R
\name{ages}
\alias{ages}
\title{Split continuous ages to age groups}
\usage{
ages(x, groups, char = FALSE)
}
\arguments{
\item{x}{Array with numeric values of age}

\item{groups}{Array with numeric values of lower boundaries of age groups}

\item{char}{Logical. Should output array be numeric (FALSE) or character (Factor) array (TRUE)}
}
\value{
Array with classified ages
}
\description{
Split continuous ages to age groups
}
\examples{
groups <- c(0,1,5,10,15,20, 24, 45, 85)
x <- sample(0:100, 101)
ages(x, groups, FALSE)

}
