% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdt.R
\name{asdt}
\alias{asdt}
\title{Associated single decrement life table (ASDT) for causes of death (cause-deleted life table)}
\usage{
asdt(age, sex, m_all, m_i, full = F, method = "chiang1968")
}
\arguments{
\item{age}{Array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{sex}{Sex. "m" for males or "f" for females.}

\item{m_all}{Age specific mortality rates of all causes of death}

\item{m_i}{Age specific mortality rates of some cause of death (i)}

\item{full}{Logical. Is full table needed? TRUE = full, FALSE = concise}

\item{method}{The method of ASDT construction to use. Now just "chiang1968" is supported.}
}
\value{
dataframe
}
\description{
Associated single decrement life table (ASDT) for causes of death (cause-deleted life table)
}
