% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccm.R
\name{ccm}
\alias{ccm}
\title{Cohort Component Model for Projecting Population}
\usage{
ccm(Mx, Fx, Ix = NULL, age.mx, age.fx, N0, ...)
}
\arguments{
\item{Mx}{Matrix or dataframe with mx, where row represents age, column represents a forecasted period. Thus, first column should be the first projected period, while each row represents age-specific mortality rate.}

\item{Fx}{Matrix or dataframe with fx, where row represents age, column represents a forecasted period. Thus, first column should be the first projected period, while each row represents age-specific fertility rate.}

\item{Ix}{Optional. Matrix or dataframe with net number of migrants, where row represents age, column represents a forecasted period. Thus, first column should be the first projected period, while each row represents net number of migrants.}

\item{age.mx}{Numeric vector. Age groups for Mx rows.}

\item{age.fx}{Numeric vector. Age groups for Fx rows.}

\item{N0}{Numeric vector. Population in the initial period by age groups.}

\item{...}{Optional. Additional arguments for \code{leslie} function.}
}
\value{
Matrix with projected population. Columns represent periods, rows represent age groups. The first column is \code{N0} (initial period).
}
\description{
Cohort Component Model for Projecting Population
}
