% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demor-package.R
\docType{package}
\name{demor-package}
\alias{demor}
\alias{demor-package}
\title{demor: Basic functions for demographic analysis}
\description{
Basic functions for demographic analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vadvu.github.io/demor/}
}

}
\author{
\strong{Maintainer}: Vadim Ustyuzhanin \email{vadim.v.ustyuzhanin@gmail.com} (\href{https://orcid.org/0000-0003-3800-1108}{ORCID})

}
\keyword{internal}
