% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edagger.R
\name{edagger}
\alias{edagger}
\title{e-dagger}
\usage{
edagger(age, mx, ...)
}
\arguments{
\item{age}{Numeric array of age intervals}

\item{mx}{Numeric array with age specific mortality rates.}

\item{...}{Optional. Additional arguments for \code{\link[=LT]{LT()}} function.}
}
\value{
Numeric array
}
\description{
e-dagger
}
