% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert.approx.R
\name{fert.approx}
\alias{fert.approx}
\title{Fertility models for ASFR approximation}
\usage{
fert.approx(fx, age, model, start = NULL, se = F, alpha = 0.05, bn = 1000)
}
\arguments{
\item{fx}{Numeric vector of age specific fertility rates.}

\item{age}{Numeric vector of ages.}

\item{model}{Character. Model name to be estimated. Now "Hadwiger", "Gamma", "Brass" and "Beta" are supported.}

\item{start}{Numeric vector with user-specific values of parameters for optimization. Default is \code{NULL} (choose automatically)}

\item{se}{Logical. Should bootstrapped variance for ASFR approximation be calculated. Default is \code{FALSE} for no bootstrap.}

\item{alpha}{Numeric. Used if \code{se = TRUE}, the level of uncertainty. By default, \code{alpha = 0.05} for 95\% CI.}

\item{bn}{Numeric. Used if \code{se = TRUE}, number of bootstrap samples. By default, \code{bn = 1000}.}
}
\value{
list with estimated model (parameters, variance-covariance matrix, percentiles of parameters) and dataframe with predicted and observed ASFR as well as SE and percentile of predictions
}
\description{
Fertility models for ASFR approximation
}
\details{
This function runs least squares optimization of the selected fertility function with 1e-06 as tolerance parameter.
\subsection{Hadwiger model}{

The model is as follows: \deqn{f(age) = \frac{ab}{c} \frac{c}{age}^{3/2} exp[-b^2(\frac{c}{age}+\frac{age}{c}-2)]}
}

\subsection{Gamma model}{

The model is as follows: \deqn{f(age) = \frac{R}{\Gamma(b)c^b}(age-d)^{b-1} exp[-(\frac{age-d}{c})]}
}

\subsection{Brass model}{

The model is as follows: \deqn{f(age) = \frac{R}{\Gamma(b)c^b}(age-d)^{b-1} exp[-(\frac{age-d}{c})]}
}
}
\examples{

# fert.approx(fx = ASFR, age = 15:55, model = "Hadwiger", se = FALSE)

}
\references{
Peristera, P., & Kostaki, A. (2007). Modeling fertility in modern populations. \emph{Demographic Research}, \emph{16}, 141-194.
}
