% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert.approx.R
\name{fert.approx}
\alias{fert.approx}
\title{Fertility models for ASFR approximation}
\usage{
fert.approx(fx, age, model, boot = F)
}
\arguments{
\item{fx}{Numeric vector of age specific fertility rates.}

\item{age}{Numeric vector of ages.}

\item{model}{Character. Model name to be estimated. Now "Hadwiger" and "Gamma" are supported.}

\item{boot}{Logical. Should bootstrapped 95\% confidence intervals for ASFR approximation be calculated. Default is \code{FALSE} for no bootstrap.}
}
\value{
list with estimated model (parameters, R-squred, variance-covariance matrix of parameters) and dataframe with predicted ASFR.
}
\description{
Fertility models for ASFR approximation
}
\details{
This function runs least squares optimization of the selected fertility function using Port algorithm with 2000 maximum iterations
and 1e-07 as tolerance parameter.
\subsection{Hadwiger model}{

The model is as follows: \deqn{f(age) = \frac{ab}{c} \frac{c}{age}^{3/2} exp[-b^2(\frac{c}{age}+\frac{age}{c}-2)]}
}

\subsection{Gamma model}{

The model is as follows: \deqn{f(age) = \frac{R}{\Gamma(b)c^b}(age-d)^{b-1} exp[-(\frac{age-d}{c})]}
}
}
\examples{

# fert.approx(fx = ASFR. age = 15:55, model = "Hadwiger", boot= F)

}
\references{
Peristera, P., & Kostaki, A. (2007). Modeling fertility in modern populations. \emph{Demographic Research}, \emph{16}, 141-194.
}
