% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hli.R
\name{hli}
\alias{hli}
\title{The Human Life Indicator (HLI)}
\usage{
hli(age, sex, mx)
}
\arguments{
\item{age}{Array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{sex}{Sex. "m" for males or "f" for females.}

\item{mx}{Age specific mortality rates.}
}
\value{
HLI value
}
\description{
The Human Life Indicator (HLI)
}
\references{
Ghislandi, S., Sanderson, W.C., & Scherbov, S. (2019). A Simple Measure of Human Development: The Human Life Indicator. \emph{Population and Development Review}, \emph{45}, 219–233.
}
