% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hli.R
\name{hli}
\alias{hli}
\title{The Human Life Indicator (HLI)}
\usage{
hli(age, mx, ...)
}
\arguments{
\item{age}{Numeric array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{mx}{Numeric array with age specific mortality rates.}

\item{...}{Optional. Additional arguments for \code{LT} function.}
}
\value{
HLI value
}
\description{
The Human Life Indicator (HLI)
}
\details{
It is calculated as
\deqn{HLI = \prod_{x=\alpha}^{\omega}(x + a_x)^d_x}
where \eqn{\alpha, \omega} are the first and last age groups, \eqn{x} is age, \eqn{a_x, d_x} are life table functions.
}
\references{
Ghislandi, S., Sanderson, W.C., & Scherbov, S. (2019). A Simple Measure of Human Development: The Human Life Indicator. \emph{Population and Development Review}, \emph{45}, 219–233.
}
