% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leecarter.R
\name{leecart}
\alias{leecart}
\title{(IN DEMO!) Basic Lee-Carter model}
\usage{
leecart(data, n = 10)
}
\arguments{
\item{data}{Dataframe with following columns: age, year, mx}

\item{n}{Numeric. Forecasted horizon}
}
\value{
Dataframe with projected mx for t+n periods with mean, low95 and high 95 values (group column)
}
\description{
(IN DEMO!) Basic Lee-Carter model
}
