% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie.R
\name{leslie}
\alias{leslie}
\title{Leslie Matrix}
\usage{
leslie(mx, fx, age1, age2, sex)
}
\arguments{
\item{mx}{Vector of age specific mortality rates.}

\item{fx}{Vector of age specific fertility rates.}

\item{age1}{Vector of ages for mx.}

\item{age2}{Vector of ages for fx.}

\item{sex}{Sex. "m" for males or "f" for females.}
}
\value{
Matrix.
}
\description{
Leslie Matrix
}
