% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT.R
\name{LT}
\alias{LT}
\title{Life table with corrected nax}
\usage{
LT(age, sex, mx)
}
\arguments{
\item{age}{Array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{sex}{Sex. "m" for males or "f" for females.}

\item{mx}{Age specific mortality rates.}
}
\value{
Matrix
}
\description{
Life table with corrected nax
}
