% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mac.R
\name{mac}
\alias{mac}
\title{Mean Age at Childbearing (MAC)}
\usage{
mac(fx, age)
}
\arguments{
\item{fx}{Numeric array of age specific fertility rates.}

\item{age}{Numeric array of ages. For example, \code{15:55} for 1-year age-groups}
}
\value{
Numeric value.
}
\description{
Mean Age at Childbearing (MAC)
}
