% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdecomp.R
\name{mdecomp}
\alias{mdecomp}
\title{Age and cause decomposition of differences in life expectancies}
\usage{
mdecomp(mx1, mx2, age, sex)
}
\arguments{
\item{mx1}{List of arrays. 1st array should be all-cause nmx in the 1st population, other arrays are cause-specific nmx in the 1st population}

\item{mx2}{List of arrays. 1st array should be all-cause nmx in the 2nd population, other arrays are cause-specific nmx in the 2nd population}

\item{age}{Array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{sex}{Sex. "m" for males and "f" for females}
}
\value{
Dataframe with 1st column as overall decomposition (\code{ex12}), and other columns are decompositions by causes (\code{ex12_cause(i)})
}
\description{
Age and cause decomposition of differences in life expectancies
}
