% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.age.R
\name{med.age}
\alias{med.age}
\title{Median age calculation}
\usage{
med.age(N, age, int = 1)
}
\arguments{
\item{N}{Array with numeric values of population by age group (from young to old)}

\item{age}{Array with numeric values age}

\item{int}{Numeric value of age interval. For example, 1 or 5}
}
\value{
Numeric value
}
\description{
Median age calculation
}
