% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.age.R
\name{med.age}
\alias{med.age}
\title{Median age calculation}
\usage{
med.age(N, age)
}
\arguments{
\item{N}{Numeric array. Population counts by age groups (from young to old)}

\item{age}{Numeric array. Lower bounds of age groups, same length as \code{N}}
}
\value{
Numeric value.
}
\description{
Median age calculation
}
