% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLT.R
\name{MLT}
\alias{MLT}
\title{Multiple Decrement Life Table}
\usage{
MLT(age, mx, ...)
}
\arguments{
\item{age}{Numeric array of age intervals; for full life table = \code{0:100}; for concise life table = \code{c(0:1, seq(5,85,5))}}

\item{mx}{List of numeric arrays. 1st array should be all-cause mx in the population, other arrays are cause-specific mx in the population.}

\item{...}{Other parameters for the function \code{LT()} including \code{ax} (by default, the middle of the interval), \code{sex} (by default = \code{"m"} - males), \code{l0} (by default = 1).}
}
\value{
Extended \code{LT} matrix
}
\description{
Multiple Decrement Life Table
}
