% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mdecomp.R
\name{plot.mdecomp}
\alias{plot.mdecomp}
\title{Plot for mdecomp function}
\usage{
\method{plot}{mdecomp}(x, return.data = FALSE, ...)
}
\arguments{
\item{x}{A result of age and cause decomposition from \code{\link[=mdecomp]{mdecomp()}}.}

\item{return.data}{Should the data be returned (\code{TRUE}) or the ggplot2 plot (default, \code{FALSE}).}

\item{...}{Ignored.}
}
\value{
ggplot2 plot
}
\description{
Plot for mdecomp function
}
\seealso{
\code{\link[=mdecomp]{mdecomp()}}
}
