% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyr.R
\name{plot_pyr}
\alias{plot_pyr}
\title{Plot population pyramid}
\usage{
plot_pyr(popm, popf, ages)
}
\arguments{
\item{popm}{Numeric vector of male population}

\item{popf}{Numeric vector of female population}

\item{ages}{Numeric vector of unique ages}
}
\value{
plot as \code{ggplot2} object
}
\description{
Plot population pyramid
}
