% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_code.R
\name{print.demo_code}
\alias{print.demo_code}
\title{S3 method for printing a \code{demo_code}}
\usage{
\method{print}{demo_code}(x, ...)
}
\description{
Prints nothing; \code{demo_code} objects should be seen and not heard.
}
\details{
If the \code{demo_code} object was created by inputting a string, we should run that code and print any output.
}
