% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosbris.codes}
\alias{rosbris.codes}
\title{Rosbris' Region codes}
\format{
\subsection{A dataframe}{

\describe{
\item{n}{number of Region (sequence number)}
\item{name}{Region name}
\item{code}{Region unique code}
}
}
}
\source{
Russian Fertility and Mortality Database. Center for Demographic Research, Moscow (Russia). Available at https://www.nes.ru/demogr-fermort-data
}
\usage{
rosbris.codes
}
\description{
Rosbris' Region codes
}
\keyword{datasets}
