% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosbris_mortality_pop_1}
\alias{rosbris_mortality_pop_1}
\title{Rosbris' Mortality and population (1-year interval)}
\format{
\subsection{A dataframe}{

\describe{
\item{year}{year}
\item{code}{number of Region}
\item{territory}{rural/urban/total}
\item{sex}{sex: both/male/femal}
\item{age}{age}
\item{mx}{ASMR}
\item{N}{Population}
\item{Dx}{Number of Deaths}
\item{name}{Region name}
}
}
}
\source{
Russian Fertility and Mortality Database. Center for Demographic Research, Moscow (Russia). Available at https://www.nes.ru/demogr-fermort-data
}
\usage{
rosbris_mortality_pop_1
}
\description{
Rosbris' Mortality and population (1-year interval)
}
\keyword{datasets}
