% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosbris_mortality_pop_5}
\alias{rosbris_mortality_pop_5}
\title{Rosbris' Moertality and population (5-year interval)}
\format{
\subsection{A dataframe}{

\describe{
\item{year}{year}
\item{code}{number of Region}
\item{territory}{rural/urban/total}
\item{sex}{sex: both/male/femal}
\item{age}{age}
\item{mx}{ASMR}
\item{N}{Population}
\item{Dx}{Number of Deaths}
\item{name}{Region name}
}
}
}
\source{
Rosbris \href{http://demogr.nes.ru/index.php/ru/demogr_indicat/data}{webpage}
}
\usage{
rosbris_mortality_pop_5
}
\description{
Rosbris' Moertality and population (5-year interval)
}
\keyword{datasets}
