% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-leslie.R
\name{summary.leslie}
\alias{summary.leslie}
\title{Leslie matrix summary}
\usage{
\method{summary}{leslie}(object, d = 1, ...)
}
\arguments{
\item{object}{A leslie matrix from \code{\link[=leslie]{leslie()}}.}

\item{d}{Time step. By default it is 1. Only affects \eqn{r}.}

\item{...}{Ignored.}
}
\description{
Leslie matrix summary
}
\details{
The function calculates \eqn{\lambda}, \eqn{r}, \eqn{w} and \eqn{v}.
\itemize{
\item \eqn{\lambda} -- asymptotic growth factor that is dominant eigenvalue.
\item \eqn{r} -- asymptotic growth rate that is \eqn{ln \lambda / \delta t}.
\item \eqn{w} -- stable age distribution normalized to 1 s.t. \eqn{\sum_x^{\omega} w_x = 1} where \eqn{x} is age.
\item \eqn{v} -- reproductive values normalized s.t. \eqn{v'w = 1}.
}
}
\seealso{
\code{\link[=leslie]{leslie()}}
}
