% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tatfr.R
\name{tatfr}
\alias{tatfr}
\title{Tempo-adjusted total fertility rate (TFR')}
\usage{
tatfr(past_fx, present_fx, post_fx, age)
}
\arguments{
\item{past_fx}{List with numeric arrays of age specific fertility rates for period t-1 by parity}

\item{present_fx}{List with numeric arrays of age specific fertility rates for period t by parity (it is period of interest)}

\item{post_fx}{List with numeric arrays of age specific fertility rates for period t+1 by parity}

\item{age}{Array with numeric values age}
}
\value{
list with TFR' (tatfr) and TFR' by parity (tatfr_i, in user-specific order as in lists), TFR (tfr) and TFR by parity (tfr_i)
}
\description{
Tempo-adjusted total fertility rate (TFR')
}
