% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr.R
\name{tfr}
\alias{tfr}
\title{Total Fertility Rate (TFR)}
\usage{
tfr(fx, age.int = 1)
}
\arguments{
\item{fx}{Numeric array of age specific fertility rates.}

\item{age.int}{Numeric. Age group: \code{1} for one-year, \code{5} for five-year. Any age groups are allowed.}
}
\value{
Numeric value.
}
\description{
Total Fertility Rate (TFR)
}
