% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yll.R
\name{yll}
\alias{yll}
\title{Years of Life Lost (YLL) calculation}
\usage{
yll(
  Dx,
  type = c("yll", "yll.p", "yll.r", "asyr"),
  Dx_all = NULL,
  pop = NULL,
  w = NULL
)
}
\arguments{
\item{Dx}{Array with the number of deaths.}

\item{type}{Character. Type of YLL to calculate. See details section.}

\item{Dx_all}{Array with the number of all deaths. Used only with \code{yll.p} type, where \code{Dx} is array with cause-specific deaths.}

\item{pop}{Array with population. Used only with \code{yll.r} and \code{asyr} types.}

\item{w}{Array with population weights for direct standardization. Used only with \code{asyr} type.}
}
\value{
list with values.
}
\description{
Years of Life Lost (YLL) calculation
}
\details{
\enumerate{
\item Absolute number of \emph{YLL} - \code{yll} type.
\item \emph{YLL} as proportion - \code{yll.p} type.
\item \emph{YLL} rate - \code{yll.r} type.
\item Age-standardized \emph{YLL} rate  - \code{asyr} type.
For more details see \href{https://github.com/vadvu/demor}{github page}.
}
}
\references{
Martinez, R., Soliz, P., Caixeta, R., Ordunez, P., 2019. Reflection on modern methods: years of life lost due to premature mortality—a versatile and comprehensive measure for monitoring non-communicable disease mortality. International Journal of Epidemiology 48, 1367–1376.
}
